Add 'infer merges true' to tag rules.
Note that this needs an unreleased version of svn2git!
The svn2git patch that makes this work is hackish and incompatible
(because it turns merge inference off by default, breaking existing rules),
so I haven't committed it yet.

The idea is to disable merge inference for branches, because it causes
unwanted merge commits for things like BOINC r25688. But we still need it
for tags coming from CVS, to figure out the correct parent commit.